resolve patches with llvm version bump
This bump to llvm happens to resolve two issues.
1. https://bugs.llvm.org/show_bug.cgi?id=34396
2. https://bugs.llvm.org/show_bug.cgi?id=34392

The workaround to (1) used to exist as a patch file in this repository.
It is no longer needed. Code relating to it has also been deleted.

The resolution to (2) used to exist as a patch in this repo, but ended
up getting pulled out and merged into WebGHC's fork of GHC as an actual
commit.